+++ backend/test/backend_tests/helpers.clj
@@ -491,6 +491,7 @@
     ([key default]
      (get data key (get cf/config key default)))))
 
+
 (defn reset-mock!
   [m]
   (swap! m (fn [m]